Move TemporaryDirectory to dmlc/filesystem.h
