+++ test/unittest/unittest_inputsplit.cc
@@ -1,4 +1,5 @@
 #include <dmlc/data.h>
+#include <dmlc/filesystem.h>
 #include <string>
 #include <iostream>
 #include <fstream>
@@ -8,16 +9,6 @@
 #include <future>
 #include <cstdlib>
 #include <gtest/gtest.h>
-
-/* platform specific headers */
-#ifdef _WIN32
-#define NOMINMAX
-#include <windows.h>
-#include <Shlwapi.h>
-#pragma comment(lib, "Shlwapi.lib")
-#else
-#include <unistd.h>
-#endif
 
 static inline void CountDimensions(dmlc::Parser<uint32_t>* parser,
                                    size_t* out_num_row, size_t* out_num_col) {
@@ -35,100 +26,11 @@ static inline void CountDimensions(dmlc::Parser<uint32_t>* parser,
   *out_num_col = num_col;
 }
 
-class TemporaryDirectory {
- public:
-  TemporaryDirectory() {
-#if _WIN32
-    /* locate the root directory of temporary area */
-    char tmproot[MAX_PATH] = {0};
-    const DWORD dw_retval = GetTempPathA(MAX_PATH, tmproot);
-    if (dw_retval > MAX_PATH || dw_retval == 0) {
-      LOG(FATAL) << "TemporaryDirectory(): "
-                 << "Could not create temporary directory";
-    }
-    /* generate a unique 8-letter alphanumeric string */
-    const std::string letters = "abcdefghijklmnopqrstuvwxyz0123456789_";
-    std::string uniqstr(8, '\0');
-    std::random_device rd;
-    std::mt19937 gen(rd());
-    std::uniform_int_distribution<int> dis(0, letters.length() - 1);
-    std::generate(uniqstr.begin(), uniqstr.end(),
-      [&dis, &gen, &letters]() -> char {
-        return letters[dis(gen)];
-      });
-    /* combine paths to get the name of the temporary directory */
-    char tmpdir[MAX_PATH] = {0};
-    PathCombineA(tmpdir, tmproot, uniqstr.c_str());
-    if (!CreateDirectoryA(tmpdir, NULL)) {
-      LOG(FATAL) << "TemporaryDirectory(): "
-                 << "Could not create temporary directory";
-    }
-    path = std::string(tmpdir);
-#else
-    std::string tmproot; /* root directory of temporary area */
-    std::string dirtemplate; /* template for temporary directory name */
-    /* Get TMPDIR env variable or fall back to /tmp/ */
-    {
-      const char* tmpenv = getenv("TMPDIR");
-      if (tmpenv) {
-        tmproot = std::string(tmpenv);
-        // strip trailing forward slashes
-        while (tmproot.length() != 0 && tmproot[tmproot.length() - 1] == '/') {
-          tmproot.resize(tmproot.length() - 1);
-        }
-      } else {
-        tmproot = "/tmp";
-      }
-    }
-    dirtemplate = tmproot + "/tmpdir.XXXXXX";
-    std::vector<char> dirtemplate_buf(dirtemplate.begin(), dirtemplate.end());
-    dirtemplate_buf.push_back('\0');
-    char* tmpdir = mkdtemp(&dirtemplate_buf[0]);
-    if (!tmpdir) {
-      LOG(FATAL) << "TemporaryDirectory(): "
-                 << "Could not create temporary directory";
-    }
-    path = std::string(tmpdir);
-#endif
-    LOG(INFO) << "Created temporary directory " << path;
-  }
-
-  ~TemporaryDirectory() {
-    for (const std::string& filename : file_list) {
-      if (std::remove(filename.c_str()) != 0) {
-        LOG(FATAL) << "Couldn't remove file " << filename;
-      }
-    }
-#if _WIN32
-    const bool rmdir_success = (RemoveDirectoryA(path.c_str()) != 0);
-#else
-    const bool rmdir_success = (rmdir(path.c_str()) == 0);
-#endif
-    if (rmdir_success) {
-      LOG(INFO) << "Successfully deleted temporary directory " << path;
-    } else {
-      LOG(FATAL) << "~TemporaryDirectory(): "
-                 << "Could not remove temporary directory " << path;
-    }
-  }
-
-  std::string AddFile(const std::string& filename) {
-    const std::string file_path = this->path + "/" + filename;
-    file_list.push_back(file_path);
-    return file_path;
-  }
-
-  std::string path;
-
- private:
-  std::vector<std::string> file_list;
-};
-
 TEST(InputSplit, test_split_csv_noeol) {
   size_t num_row, num_col;
   {
     /* Create a test case for partitioned csv with NOEOL */
-    TemporaryDirectory tempdir;
+    dmlc::TemporaryDirectory tempdir;
     {
       const std::string filename = tempdir.AddFile("train_0.csv");
       std::ofstream of(filename.c_str(), std::ios::binary);
@@ -159,7 +61,7 @@ TEST(InputSplit, test_split_csv_noeol) {
 TEST(InputSplit, test_split_libsvm_noeol) {
   {
     /* Create a test case for partitioned libsvm with NOEOL */
-    TemporaryDirectory tempdir;
+    dmlc::TemporaryDirectory tempdir;
     const char* line
       = "1 3:1 10:1 11:1 21:1 30:1 34:1 36:1 40:1 41:1 53:1 58:1 65:1 69:1 "
         "77:1 86:1 88:1 92:1 95:1 102:1 105:1 117:1 124:1";
@@ -192,7 +94,7 @@ TEST(InputSplit, test_split_libsvm) {
   size_t num_row, num_col;
   {
     /* Create a test case for partitioned libsvm */
-    TemporaryDirectory tempdir;
+    dmlc::TemporaryDirectory tempdir;
     const int nfile = 5;
     for (int file_id = 0; file_id < nfile; ++file_id) {
       const std::string filename
@@ -216,7 +118,7 @@ TEST(InputSplit, test_split_libsvm) {
 TEST(InputSplit, test_split_libsvm_distributed) {
   {
     /* Create a test case for partitioned libsvm */
-    TemporaryDirectory tempdir;
+    dmlc::TemporaryDirectory tempdir;
     const char* line
       = "1 3:1 10:1 11:1 21:1 30:1 34:1 36:1 40:1 41:1 53:1 58:1 65:1 69:1 "
         "77:1 86:1 88:1 92:1 95:1 102:1 105:1 117:1 124:1\n";
